fimal updates

+++ API/dataBase/models/productModel.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(20, 2),
         allowNull: false,
       },
       color: {
